backend/requirements.txt
Flask
Flask-Session
Flask-Migrate
Flask-SQLAlchemy
requests
gunicorn
alembic
mysqlclient
python-dotenv